requirements.txt
fastapi==0.110.3
uvicorn==0.33.0
python-multipart==0.0.20
pydantic==2.8.2
pydantic-settings==2.5.2
scikit-learn==1.3.2
numpy==1.24.4
PyPDF2==3.0.1
google-genai==1.2.0; python_version >= "3.9"
